fix: move toast provider from admin (#105)
