--- src/app/admin/providers/toast-client.provider.ts
@@ -1,2 +1,0 @@
-"use client";
-export { ToastProvider } from "../providers/toast.provider";
